% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_test_or.R
\name{af_test_or}
\alias{af_test_or}
\title{Estimate odds ratios (ORs) between two conditions}
\usage{
af_test_or(x1, n1, x2, n2, level = 0.95)
}
\arguments{
\item{x1}{The number of alleles of one type in the first dataset.}

\item{n1}{The total numbers of alleles (of all types) in the first dataset.}

\item{x2}{The number of alleles of one type in the second dataset.}

\item{n2}{The total numbers of alleles (of all types) in the second dataset.}

\item{level}{The level for the confidence interval.}
}
\value{
A tibble with rows matching inputs and with these columns:
\itemize{
\item \code{OR}: the estimated odds ratio.
\item \code{CIL}: lower confidence interval.
\item \code{CIU}: upper confidence interval.
\item \code{pval}: the p-value.
}
}
\description{
This function tests if the ORs between allele frequencies at each locus is non-zero between two datasets based on their counts and sample sizes, returning ORs (odds of first condition over odds of second), confidence intervals, and p-values.
Under the null hypothesis, the OR is 1 in both datasets, while under the alternative it is non-zero.
Counts are assumed to be Binomially distributed.
If inputs are matrices of equal dimensions, rows are tested separately but columns are combined, assigned a single OR assuming that it is shared across columns (useful for combining data from different ancestries, or different loci).
Vector inputs are treated as column matrices.
Parameter fitting and significance is calculated using \code{\link[stats:glm]{stats::glm()}}, with dataset as the effect variable and column (when more than one) as a categorical covariate conditioned upon.
}
\details{
This function accepts exactly the same inputs as \code{\link[=af_test]{af_test()}}, which it complements by providing odds ratios and 95\% confidence intervals, which the former does not return at all.
However, a shared OR across columns of the same row/locus requires a more restricted alternative hypothesis with one alternative parameter, the OR, and thus a 1 degree of freedom test, so these p-values are different from those of \code{\link[=af_test]{af_test()}}.
(In contrast, \code{\link[=af_test]{af_test()}} implicitly allows ORs to be different for different columns, and they can even go in opposite directions, so the test has degrees of freedom equal to the number of columns.)
}
\examples{
### Test single locus ###

# Simulate data where the alternative hypothesis is true
# true allele frequencies are different
p1 <- 0.2
p2 <- 0.3
# sample sizes can be very different
n1 <- 100
n2 <- 50
# simulate counts from Binomial distribution
x1 <- rbinom( 1, n1, p1 )
x2 <- rbinom( 1, n2, p2 )

# this performs desired statistical test
data <- af_test_or( x1, n1, x2, n2 )
# inspect stats!
# though null hypothesis is true, there is low power because sample sizes are small,
# so example is most likely insignificant
data

### Multiple loci and ancestries ###

# it's useful to think of each column as the same locus but a different ancestry,
# so the allele frequencies can be different.
# Each row is a different locus to be tested separately.
# here the null hypothesis is true so both datasets will share the same allele frequency.
# will simulate 2 loci, 3 ancestries
p <- matrix(
    c(
        0.2, 0.7, 0.5,
        0.1, 0.3, 0.5
    ),
    nrow = 2,
    ncol = 3,
    byrow = TRUE
)
# all sample sizes can be different
n1 <- matrix( c( 100, 80, 200, 33, 66, 99 ), nrow = 2, byrow = TRUE )
n2 <- matrix( c( 50, 15, 99, 44, 55, 66 ), nrow = 2, byrow = TRUE )
# simulate counts from Binomial distribution (vectorized across all 3 ancestries,
# reshaped as matrix of desired dimensions)
x1 <- matrix( rbinom( 6, n1, p ), nrow = 2 )
x2 <- matrix( rbinom( 6, n2, p ), nrow = 2 )

# perform desired statistical test, a single p-value testing whether the shared OR across 
# ancestries differs from 1.
data <- af_test_or( x1, n1, x2, n2 )
data

}
\seealso{
\code{\link[=af_test]{af_test()}}, which unlike this function calculates p-values (but not ORs) under a more general alternative hypothesis that the OR is non-zero with different values in each column.
}
