% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_geno.R
\name{bias_geno}
\alias{bias_geno}
\title{Simulate genotyping platform bias}
\usage{
bias_geno(X, p = NA, biased_loci = NULL, shape1 = 1, shape2 = 5)
}
\arguments{
\item{X}{True genotype matrix, with values in 0, 1, or 2 counting reference alleles.
This matrix must have loci along the rows and individuals along columns.}

\item{p}{Proportion of biased loci.
Either this or \code{biased_loci} must be provided.
If provided and \code{biased_loci} is \code{NULL}, then biased loci will be picked randomly from among all loci.}

\item{biased_loci}{The indexes of loci desired to be biased.
Either this or \code{p} must be provided.
If provided, this is the set of loci that will be biased, and \code{p} is ignored.}

\item{shape1}{The first shape parameter of the Beta distribution used to generate error probabilities per locus.}

\item{shape2}{The second shape parameter of the Beta distribution used to generate error probabilities per locus.}
}
\value{
A named list with the following elements
\itemize{
\item \code{X}: The modified genotype matrix, with the same dimensions as the input \code{X}.  A proportion \code{p} of its loci will have platform biases, the rest of the loci are unaltered.
\item \code{biased_loci}: The length \code{m_biased} vector of indexes of biased loci.
\item \code{error_probs}: The length \code{m_biased} vector of randomly drawn error probabilities from the Beta distribution, one per locus.
\item \code{error_genos}: The length \code{m_biased} vector of randomly drawn erroneous genotypes, one per locus.
}
}
\description{
This function accepts a true genotype matrix as the input and either the proportion of biased loci or their exact indexes, and simulates biased genotypes determined by those and some additional parameters.
The goal is to simulate genotyping platform-specific errors where there is a greater tendency to read a certain allele over the other.
In particular, once biased loci are determined, each of those loci is drawn from a Beta distribution a random probability that individuals at that locus will have an erroneous genotype (\code{error_prob}), and a value for the erroneous genotype (\code{error_geno}, either 0 or 2, picked with equal probability).
Then, at the same biased locus, a proportion \code{error_prob} of random individuals have their genotypes replaced by the \code{error_geno} value, which therefore results in biased genotypes towards the value of \code{error_geno}.
}
\examples{
# simulate a toy genotype matrix with 10 loci and 3 individuals
X <- matrix(
    rbinom( 30, 2, 0.5 ),
    nrow = 10,
    ncol = 3
)

# simulate platform bias!
obj <- bias_geno( X, p = 0.1 )

# the genotype matrix with platform-specific biases is here
X_biased <- obj$X

# the other elements in the list `obj` provide further details
# regarding how the platform biases were generated

}
