% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_category.R
\name{filter_category}
\alias{filter_category}
\title{Calculate locus categories depending on threshold}
\usage{
filter_category(data, pcut)
}
\arguments{
\item{data}{A tibble/data.frame with at least three columns:
\itemize{
\item \code{revcomp}: Logical, whether locus is identified as having reverse complement (flippale) alleles.
\item \code{pval_fwd}: p-values in given orientation.
\item \code{pval_rev}: p-values in reversed orientation for the second dataset, for loci with \code{revcomp == TRUE}.  Values for rest of loci are ignored.
}}

\item{pcut}{The p-value threshold}
}
\value{
A vector with length equal to the number of rows of \code{data}, with one of 3 possible values depending on classification outcome per locus: "remove", "flip", or "keep".
}
\description{
This takes the output of \code{\link[=af_filter]{af_filter()}} or an equivalent function and classifies loci depending on the pattern of significance, which is threshold-dependent.
For loci that only have forward p-values, the are classified as "remove" if they are significant (below the threshold) or "keep" otherwise.
For loci with both forward and reversed p-values (which must have reversed complement ref/alt alleles), they are classified as "remove" if both p-values are significant, "flip" if the forward p-value is the smallest, and "keep" otherwise.
}
\examples{
library(tibble)
# an example with all cases
data <- tibble(
    revcomp  = c( FALSE, FALSE,  TRUE,  TRUE,  TRUE,  TRUE,  TRUE ),
    pval_fwd = c(   0.9, 1e-10,   0.1,   0.9,   0.9, 1e-10, 1e-10 ),
    pval_rev = c(    NA,    NA,   0.9,   0.1, 1e-10,   0.9, 1e-10 )
)
pcut <- 0.01

# get categories!
cat <- filter_category( data, pcut )

}
\seealso{
\code{\link[=af_filter]{af_filter()}}
}
