% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_revcomps.R
\name{flip_revcomps}
\alias{flip_revcomps}
\title{Simulate strand flipping for ambiguous ref/alt pairs}
\usage{
flip_revcomps(X, p = NA, bim = NULL, flipped_loci = NULL)
}
\arguments{
\item{X}{True genotype matrix, with values in 0, 1, or 2 counting reference alleles.
This matrix must have loci along the rows and individuals along columns.}

\item{p}{Proportion of flippable loci to flip.
Either this and \code{bim} or \code{flipped_loci} must be provided.
If provided and \code{flipped_loci} is \code{NULL}, then flipped loci will be picked randomly from among all flippable loci.}

\item{bim}{A data frame or tibble with \code{ref} and \code{alt} allele columns for every locus in \code{X}, used only to determine which loci are flippable.
Either this and \code{p} or \code{flipped_loci} must be provided.}

\item{flipped_loci}{The indexes of loci desired to be flipped.
Either this or \code{p} and \code{bim} must be provided.
If provided, this is the set of loci that will be flipped, and \code{p} and \code{bim} are ignored.
In this case \code{flipped_loci} is not checked against the list of flippable loci.}
}
\value{
A named list with the following elements
\itemize{
\item \code{X}: The modified genotype matrix, with the same dimensions as the input \code{X}.  A proportion \code{p} of its flippable loci will be flipped, the rest of the loci are unaltered.
\item \code{flipped_loci}: The vector of indexes of flipped loci.
}
}
\description{
This function transforms a genotype matrix by "flipping" some of its loci so that the opposite allele is counted.
In other words, its numeric value \code{x}, which should be in 0, 1, or 2, is replaced with \code{2-x} but only at some proportion \code{p} of loci that are flippable.
Flippable loci are those where the alt allele is the reverse complement of the ref allele, so that there is strand ambiguity and alleles cannot be aligned trivially to external datasets.
}
\examples{
# simulate a toy genotype matrix with 5 loci and 3 individuals
X <- matrix(
    rbinom( 15, 2, 0.5 ),
    nrow = 5,
    ncol = 3
)

# for our fake BIM table, only the first three are flippable (reverse complement ref/alt)
library(tibble)
bim <- tibble(
    alt = c('A', 'G', 'C', 'T', 'GATTACA'),
    ref = c('T', 'C', 'G', 'G', 'G')
) 

# randomly pick two of the three flippable loci to flip!
obj <- flip_revcomps( X, p = 2/3, bim )

# this is the modified genotype matrix
obj$X

# and the indexes of the flipped loci
obj$flipped_loci

}
\seealso{
\code{\link[=revcomp]{revcomp()}} for reverse complements.
}
