\name{fst}
\Rdversion{0.8}
\alias{fst}
\title{Compute the \eqn{F_{ST}} for a Sample of Genes}
\description{
Compute the \eqn{F_{ST}} for a sample of genes.
}
\usage{
  fst(sample)
}
\arguments{
  \item{sample}{an object generated by the \code{\link{sim.inference.model}} or by the \code{\link{sim.coalescent}} commands}
}
\details{
Once the \code{\link{sim.inference.model}} or by the \code{\link{sim.coalescent}} command lines have been executed, \code{\link{fst}} can be used to compute the Weir and Cockerham's (1984) estimate of \eqn{F_{ST}} for that sample. The multilocus estimator of \eqn{F_{ST}} is computed as in Genepop (Rousset 2008).
}
\value{
	Weir and Cockerham's (1984) estimate of \eqn{F_{ST}} for the sample of genes contained in the object \sQuote{sample}
}
\examples{
## This is to simulate a sample of genes (at a single locus), using the inference model, with
## 50 genes collected in each of 10 sampled demes. In this example, the product of
## twice the effective population size and migration rate is 2, 
## and the frequency of allele A in the migrant pool is 0.5

sample <- sim.inference.model(number.of.sampled.demes = 10,sample.sizes = 50,M = 2,pi = 0.5)

## This is to compute Weir and Cockerham's (1984) estimate for that sample

fst(sample)
}
\references{
Rousset, F. (2008) genepop'007: a complete re-implementation of the genepop software for Windows and Linux, \emph{Molecular Ecology Notes} \bold{8}: 103-106.

Weir, B. S. and Cockerham, C. C. (1984) Estimating \emph{F}-statistics for the analysis of population structure, \emph{Evolution} \bold{38}: 1358-1370.
}