% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_eval.R
\name{gwas_eval}
\alias{gwas_eval}
\title{Calculate GWAS-inspired classification statistics for flip/remove/keep categories}
\usage{
gwas_eval(true, pred, names = TRUE)
}
\arguments{
\item{true}{Vector of true string categories, with values in "flip", "remove", or "keep".}

\item{pred}{Vector of predicted string categories, with values in "flip", "remove", or "keep".}

\item{names}{Logical for whether the return vector should have names or not.}
}
\value{
A named vector (unless \code{names = FALSE}) with five values in this order:
\itemize{
\item \code{precision}: \code{TP / ( TP + FP )}
\item \code{recall}: \code{TP / ( TP + FN )}
\item \code{TP}: the number of rows where both \code{true} and \code{pred} are \code{TRUE}
\item \code{FP}: the number of rows where \code{true} is \code{FALSE} and \code{pred} is \code{TRUE}
\item \code{FN}: the number of rows where \code{true} is \code{TRUE} and \code{pred} is \code{FALSE}
}
}
\description{
This function calculates precision, recall, and the underlying true positive (TP), false positive (FP), and false negative (FN) counts from two vectors whose values are "flip", "remove" or "keep", using a special criterion that is specific to what occurs in these cases when biased or flipped loci are tested in a GWAS, differentiating between removals and the rest and whether non-removals are correct or not.
In particular, here \code{TP} counts the number of keep and flip that are correctly predicted; \code{FP} is the number of predicted keep or flip that are incorrect; \code{FN} is the number of predicted remove that are incorrect; lastly, the implied \code{TN} (not explicitly calculated) is the number of correct removals.
}
\details{
Categories must match "keep", "flip" and "remove" in spelling and capitalization, otherwise an error is thrown.
}
\examples{
# the input data should look something like this:
true <- c('keep', 'keep',   'keep', 'flip',   'flip', 'remove', 'remove', 'remove')
pred <- c('keep', 'flip', 'remove', 'flip', 'remove', 'remove',   'flip',   'keep')

# get stats!
data <- gwas_eval( true, pred )
data

}
\seealso{
\code{\link[=binary_eval]{binary_eval()}}, which is used to calculate the desired statistics for logical (binary) classification.

\code{\link[=filter_eval]{filter_eval()}} for a wrapper applying \code{\link[=gwas_eval]{gwas_eval()}} and variants of \code{\link[=binary_eval]{binary_eval()}} to 3-category classifications (remove, flip, keep).
}
