\name{maximum.likelihood}
\Rdversion{0.8}
\alias{maximum.likelihood}
\title{Perform a Maximum-Likelihood Analysis of a Sample of Genes in a Subdivided Population}
\description{
Perform a maximum-likelihood analysis of a sample of genes in a subdivided population.
}
\usage{
  maximum.likelihood(sample,alpha,M,pi,graphics,true.M,true.pi)
}
\arguments{
  \item{sample}{an object generated by the \code{\link{sim.inference.model}} or by the \code{\link{sim.coalescent}} command}
  \item{alpha}{the alpha-level, which takes the default value alpha = 0.05}
  \item{M}{the support for M}
  \item{pi}{the support for pi}
  \item{graphics}{a logical variable, which is TRUE if the user wants graphics to be plotted}
  \item{true.M}{true (simulated) value of M}
  \item{true.pi}{true (simulated) value of pi}
}
\details{
Once the \code{\link{sim.inference.model}} or by the \code{\link{sim.coalescent}} command lines have been executed, \code{\link{maximum.likelihood}} can be used to compute the maximum likelihood analysis of the sample of genes.
}
\value{
	a \sQuote{sample} object
}
\examples{
## This is to simulate a sample of genes (at a single locus), using the inference model, with
## 50 genes collected in each of 10 sampled demes. In this example, the product of
## twice the effective population size and migration rate is 2, 
## and the frequency of allele A in the migrant pool is 0.5

sample <- sim.inference.model(number.of.sampled.demes = 10,sample.sizes = 50,M = 2,pi = 0.5)

## This is to compute Nei's unbiased heterozygosity for that sample

maximum.likelihood(sample)
}
