\name{PopGenInfeR}
\Rdversion{0.8}
\alias{popgen.infeR}
\title{A Pedagogical Tool to Teach the Basics of Inference in Population Genetics}
\description{This package is intended to provide pedagogical tools to teach the basics inference in population genetics. It enables to simulate data in an island model of population structure, and to estimate the model parameters by means of maximum likelihood and Markov chain Monte Carlo (MCMC) algorithms.}
\author{Renaud Vitalis \email{renaud.vitalis@inrae.fr}}
