% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcomp.R
\name{revcomp}
\alias{revcomp}
\title{Returns reverse complements of DNA variants}
\usage{
revcomp(x)
}
\arguments{
\item{x}{A vector of DNA variants}
}
\value{
A vector of the same length where every string has been reversed-complemented.
}
\description{
This is a very simple, standalone implementation that covers DNA cases: ACGT lowercase and uppercase, one or multiple letters.
Letters that do not fall in those cases are not complemented, but all strings are reversed.
}
\examples{
revcomp( c( 'GATTACA', 'TATA' ) )

}
