\name{run.mcmc}
\Rdversion{0.8}
\alias{run.mcmc}
\title{Perform a Bayesian Analysis of a Sample of Genes in a Subdivided Population}
\description{
Perform a Bayesian analysis of a sample of genes in a subdivided population.
}
\usage{
  run.mcmc(sample,chain.length,burnin,range.M,delta.M,delta.pi,alpha,graphics,
    true.M,true.pi)
}
\arguments{
  \item{sample}{an object generated by the \code{\link{sim.inference.model}} or by the \code{\link{sim.coalescent}} command}
  \item{chain.length}{total length of the chain}
  \item{burnin}{number of initial steps to discard from the chain}
  \item{range.M}{the support for M}
  \item{delta.M}{width of the uniform proposal for the parameter M}
  \item{delta.pi}{width of the uniform proposals for the parameters pi}
  \item{alpha}{the alpha-level, which takes the default value alpha = 0.05}
  \item{graphics}{a logical variable, which is TRUE if the user wants graphics to be plotted}
  \item{true.M}{true (simulated) value of M}
  \item{true.pi}{true (simulated) value of pi}
}
\details{
Once the \code{\link{sim.inference.model}} or by the \code{\link{sim.coalescent}} command lines have been executed, \code{\link{run.mcmc}} can be used to compute a Bayesian analysis of the sample of genes.
}
\value{
	A MCMC chain.
}
\examples{
## This is to simulate a sample of genes (at a single locus), using the inference model, with
## 50 genes collected in each of 10 sampled demes. In this example, the product of
## twice the effective population size and migration rate is 2, 
## and the frequency of allele A in the migrant pool is 0.5

sample <- sim.inference.model(number.of.sampled.demes = 10,sample.sizes = 50,M = 2,pi = 0.5)

## This is to run a MCMC for that sample

run.mcmc(sample)
}
